requirements.txt
PyYAML
six